subsume bsearch.h into sortuse64.c
