+++ src/sortuse64.c
@@ -14,7 +14,6 @@
 #include <R_ext/Arith.h> // NA_INTEGER, NA_REAL
 #include <R_ext/Memory.h> // R_alloc
 
-#include "bsearch.h"
 #include "integer64.h"
 #include "sort64.h"
 #include "sortuse64.h"
@@ -34,6 +33,98 @@ static inline void bitflags_set(ValueT *bitflags, IndexT k) {
 
 static inline bool bitflags_test(const ValueT *bitflags, IndexT k) {
   return (bitflags[k / 64] & (RIGHTBIT_INTEGER64 << (k % 64))) != 0;
+}
+
+static inline IndexT integer64_bsearch_asc_EQ(const ValueT *data, IndexT l, IndexT r, ValueT value) {
+  while (l < r) {
+    IndexT m = l + ((r - l) / 2);
+    if (data[m] < value)
+      l = m + 1;
+    else
+      r = m;
+  }
+  if (value < data[l] || data[l] < value)
+    return -1;
+  return l;
+}
+
+static inline IndexT integer64_lsearch_asc_GE(const ValueT *data, IndexT l, IndexT r, ValueT value) {
+  IndexT m, g, d = 1;
+  while (l < r) {
+    g = l - 1 + d;
+    m = l + ((r - l) / 2);
+    if (g < m) {
+      if (data[g] < value) {
+        l = g + 1;
+        d *= 2;
+      } else {
+        r = g;
+        break;
+      }
+    } else {
+      if (data[m] < value)
+        l = m + 1;
+      else
+        r = m;
+      break;
+    }
+  }
+  while (l < r) {
+    m = l + ((r - l) / 2);
+    if (data[m] < value)
+      l = m + 1;
+    else
+      r = m;
+  }
+  if (data[l] < value)
+    return r + 1;
+  return l;
+}
+
+static inline IndexT integer64_bosearch_asc_EQ(const ValueT *data, const IndexT *index, IndexT l, IndexT r, ValueT value) {
+  while (l < r) {
+    IndexT m = l + ((r - l) / 2);
+    if (data[index[m]] < value)
+      l = m + 1;
+    else
+      r = m;
+  }
+  if (value < data[index[l]] || data[index[l]] < value)
+    return -1;
+  return l;
+}
+
+static inline IndexT integer64_losearch_asc_GE(const ValueT *data, const IndexT *index, IndexT l, IndexT r, ValueT value) {
+  IndexT m, g, d = 1;
+  while (l < r) {
+    g = l - 1 + d;
+    m = l + ((r - l) / 2);
+    if (g < m) {
+      if (data[index[g]] < value) {
+        l = g + 1;
+        d *= 2;
+      } else {
+        r = g;
+        break;
+      }
+    } else {
+      if (data[index[m]] < value)
+        l = m + 1;
+      else
+        r = m;
+      break;
+    }
+  }
+  while (l < r) {
+    m = l + ((r - l) / 2);
+    if (data[index[m]] < value)
+      l = m + 1;
+    else
+      r = m;
+  }
+  if (data[index[l]] < value)
+    return r + 1;
+  return l;
 }
 
 SEXP r_ram_integer64_nacount(
